fixed UI issues, tried implementing db moving back to localStorage

+++ src/App.tsx
@@ -20,46 +20,52 @@ function App() {
     const data = await res.data;
     return data;
   }
+
   
-  localStorage.setItem("Offers",JSON.stringify(Offers));
-  localStorage.setItem("Like",JSON.stringify(Like));
-  localStorage.setItem("Cart",JSON.stringify(Cart));
-  localStorage.setItem("Free",JSON.stringify(Free));
-  localStorage.setItem("Inventory",JSON.stringify(Inventory));
+  // localStorage.setItem("Offers",JSON.stringify(Offers));
+  // localStorage.setItem("Like",JSON.stringify(Like));
+  // localStorage.setItem("Cart",JSON.stringify(Cart));
+  // localStorage.setItem("Free",JSON.stringify(Free));
+  // localStorage.setItem("Inventory",JSON.stringify(Inventory));
 
   useEffect(function(){
     console.log("Setting localStorage");
-    try{
-      localStorage.clear();
-      localStorage.setItem("Offers",JSON.stringify(Offers));
-      localStorage.setItem("Like",JSON.stringify(Like));
-      localStorage.setItem("Cart",JSON.stringify(Cart));
-      localStorage.setItem("Free",JSON.stringify(Free)); //I am assuming whatever Items are offered free, are not part of main inventory a different stock is maintained for them, so whenevr an item is added or removed from Free its quantity will not be reduced/increased from inventory
-      localStorage.setItem("Inventory",JSON.stringify(Inventory));
-      callDB().then(res=>{
-        // const inventoryObj:Record<string, number> = {};
-        for(let i=0; i<res.length;i++){
-          const cartItems = JSON.parse(localStorage.getItem("Cart") as string)["items"];
-          if(cartItems[res[i].name]){
-            Inventory[`${res[i].name}`] = res[i].available - cartItems[res[i].name]
+    if(!localStorage.getItem("Initialised")){
+      try{
+        localStorage.clear();
+        localStorage.setItem("Offers",JSON.stringify(Offers));
+        localStorage.setItem("Like",JSON.stringify(Like));
+        localStorage.setItem("Cart",JSON.stringify(Cart));
+        localStorage.setItem("Free",JSON.stringify(Free)); //I am assuming whatever Items are offered free, are not part of main inventory a different stock is maintained for them, so whenevr an item is added or removed from Free its quantity will not be reduced/increased from inventory
+        localStorage.setItem("Inventory",JSON.stringify(Inventory));
+        
+        callDB().then(res=>{
+          // const inventoryObj:Record<string, number> = {};
+          for(let i=0; i<res.length;i++){
+            const cartItems = JSON.parse(localStorage.getItem("Cart") as string)["items"];
+            if(cartItems[res[i].name]){
+              Inventory[`${res[i].name}`] = res[i].available - cartItems[res[i].name]
+            }
+            else{
+              Inventory[`${res[i].name}`] = res[i].available
+            }
           }
-          else{
-            Inventory[`${res[i].name}`] = res[i].available
-          }
-        }
-        // console.log(Inventory)
-        localStorage.setItem("Inventory",JSON.stringify(Inventory))
-      });
-
-    }
-    catch(E){
-      console.log(E);
+          // console.log(Inventory)
+          localStorage.setItem("Inventory",JSON.stringify(Inventory))
+          localStorage.setItem("Initialised", "true");
+        });
+  
+      }
+      catch(E){
+        console.log(E);
+      }
     }
 
     // return function(){
     //   localStorage.clear();
     // }
   },[])
+  
 
   return (
     <div className="px-20 py-20">
